Added DELETE admin group handler, started PUT.
Need to figure out what to do with the PUT handler, as if a person is deleted, then what
happens? Maybe just remove all their expenses and everything that is assigned to them?

Should they be able to leave if they are not on zero?

Admins should have more power but it is still not very clear